Use qalc instead of bc as calculator
I've discovered qalc, which does a better job at being a quick-and-dirty
calculator than bc. Floating-point operations are supported out of the
box, but the killer feature is the extensive conversion mechanic, e.g.:

> 12 in to cm

  12 * inch = 30.48 cm

I just can't figure out how to make the currency conversions work. It's
complaining that the API it contacts is returning 404... Welp. I'll
probably have to wait for an update from the devs to switch to a working
API.