Add i3 workspace for games
Workspace 4 has been repurposed for games, with the misc workspace being
bumped to workspace 5.

I've been driven to this because I spend a lot of time in the misc
workspace playing NetHack on my cool-retro-term and more recently Battle
of Wesnoth, but it doesn't fit to put my Emacs there as well.

I'm not *actually* doing work in Emacs so I'm hesitant to put it in my
general terminal workspace (workspace 1), but I also don't want to keep
it with my games in the misc workspace. So, my games get their own
workspace.

A nice side effect is that workspaces 1-5 now all have pretty icons, so
I'm not greeted with an ugly "5" when I fiddle and quickly switch
between workspaces with my left hand on the number row.