Use darker color 15
Color 15, bright white serves as my fg gray. I toned it down a bit
because the brightness was hurting my eyes.

My ZSH prompt theme is supposed to meld everything into the background
save for the current directory name and important info such as Git
branch or Node.js version; with the previous color 15, a lot of the
elements were almost as bright as normal text instead of being melded
into the background.