Change rofi bindsym to Super+Space
This is in concert with an update to my tmux config.

The Spacebar is the largest key on my keyboard, so I want to make full
use of it. I've already made Space my Vim leader key, with Control+Space
a custom, Insert-mode leader key of sorts.

Recently I've updated tmux to use Meta+Space as my prefix, which fits in
well with the idea that the Spacebar is the gateway to most useful
keybinds.

To fit in with this, I've moved the rofi bindsym to Super+Space. This
makes sense on a lot of levels:

- Opening a program with rofi is likely the most important activity I
  have to trigger with a keybind (outside of flipping across workspaces
  and windows, which are better fit for Super+1-9 and Super+hjkl,
  respectively). The previous behavior (toggle across tiling and
  floating) is something more infrequently used and doesn't deserve to
  be on such a prominent key.

- It's in keeping with the idea that modifiers are layered: Control is
  reserved for terminal programs (like Vim), Meta is reserved for tmux
  (with the exception of movement keybinds like M-b and M-f), Super is
  reserved for the desktop environment.

- Super+Space mirrors the Command+Space keybind on OSX to open Spotlight
  Search, which does something similar to what my rofi bindsym does.