Check in mutt config
I've put this off for the longest time, because I was afraid of the
security implications. I think I've already managed that well enough.

Re: sensitive config, none of the checked-in files have anything that I
don't mean to be public. The sensitive variables should be set in a
PGP-encrypted ~/.mutt/passwd.gpg file, which isn't checked in despite
being encrypted for other reasons.

~/.mutt/aliases isn't checked in, since I can't speak for any of my
contacts whether or not they'd like their email addresses in a public
repo.

Re: non-sensitive config, it's just a matter of some additional settings
and maps that I find convenient.