Use ical instead of calcurse
Instead of a full-blown calendar program which takes a lot of time to
load (for features I don't even use), I now rely on a simple wrapper
script for cal(1) to make it convenient for interactive use.

Now it's *much* faster to get from my pressing Super+c to displaying the
actual calendar I want.