Remove mouse wheel behavior in i3bar
Every once in a while I accidentally scroll the mouse with the cursor
hovering over i3bar, e.g. when I want to scroll a webpage with my
touchpad right after using the keyboard, so the mouse is placed
generally out of the way.

This causes i3 to scroll through open workspaces, which is very jarring
for me and not at all a feature I want to take advantage of. I simply
bind the mouse wheel actions to noop to get rid of this.