Make i3 workspace layout tabbed by default
Whenever I make a second window in a workspace for the first time, I
almost always tab the layout right after. It seems that I almost always
want tabbed layouts, and when I *do* want a tiled setup it's always an
explicit realization.

I simply configured i3 to be in line with this.