Tweak Selected color for tuir
I've long had a problem reading selected text in rtv/tuir, because
having the body of a post be yellow is too visually "loud" for me and
distracts me from the actual content.

For some reason, it only just now dawned on me that I could just change
the color... *facepalm*. I now have the foreground back to the default
white, with just a gray background to really convey a cursor-like feel
to selected posts.