Use flameshot for selected-area screenshots
flameshot gui is much nicer than scrot -s:

- more featureful saving options (Imgur, clipboard, file)
- less buggy, resizeable area drag
- editing features built-in

The only downside is it seems to act more like a macOS application, in
that it sticks around after starting once and displays a tray icon.
Perhaps it's configurable to quit after a screenshot -- I don't want to
simply hide the tray icon if the process still lingers afterwards -- but
that can be a problem for another time.