Tweak tig commit display
Colors have been reduced across the board in tig's commit display (when
you press <CR> on a commit in the main view -- technically called diff
view).

A lot of metadata are muted into the background with grays, so it's much
easier to scan the commit at first glance, while at the same time
retaining the information on the screen if I need a closer look.

I've also tweaked my log date format, because the default one
infuriatingly puts the year *after* the time string.