Add Phys.org newsboat URLs
I've come to like Phys.org! Finally, a good place to get physics news
from. Previously, all the physics-related content I consumed was from
Physics SE, as well as the occasional article submitted to HN, which
unfortunately gets quite a bit of discussion with all the confidence
(dare I say, arrogance?) of the typical HN commenter, with none of the
usual expertise.