Make aerc archive keybind general
I've discovered that you can set the archive folder for accounts in
accounts.conf! That means I can set that to the Gmail-specific
"[Gmail]/All Mail" for my Gmail accounts and :archive will Do the Right
Thing™.

I've reverted my A keybind to make use of :archive now, in light of
this.