Revamp nethackrc
Biggest additions would be the colors:

- menucolors: I've added color niceties to quickly see important BUC
  status in the inventory screen.
- statushilites: I've added colors for important conditions as well as
  newly changed stats, especially stuff like HP and power/mana.

While I've added a lot more color than is default, I've made sure to
still only stick to my "highlight only when exceptional" mantra, so as
not to overload the UI with color for me and defeat the whole purpose.

Because of the huge additions, I've added some sectioning and nice
comments to the file.