Use locktoggle state to initialize DPMS
locktoggle uses a simple lockfile to tell whether DPMS is toggled on or
off -- I've had my xinitrc read this for years, but only on my work
machine. I'm checking it in even though I think it's a bit brittle
because I haven't thought of a better solution in the years since, and
it's never broken on me tbh.