Add keybind for Compose
I use my right modifiers way too much to be happy with an asymmetrical
Compose key, so I bind it to Super+Alt+u instead. This is a bit more
cumbersome (e.g. I have to do Super+Alt+u, e, " as opposed to just
doing RightAlt+e, "), so this is still something that may change as I
test it out.

Honestly been testing it out for a few months already; I'm only just now
checking it in.