Remove lockscreen on DPMS
Because of WFH, I've found that there's so much more useless friction
for me retyping my password every time my screen dims, versus the times
when I'm outside and it actually helps.

I've trialed this for a bit and I'm really happy with the convenience.
Time will tell if I remember to turn this back on when I'm outside, but
I'm usually generally conscious enough to explicitly lock my screen via
a keybind. We'll see!